VPR CTD data

A dataframe including all CTD parameters from the VPR CTD, produced by vpr_ctd_read

ctd_dat_combine

Format

A dataframe with 15 variables

time_ms

Time stamp when ROI was collected (milliseconds)

conductivity

Conductivity collected by the VPR CTD

pressure

Pressure measured from the VPR CTD in decibars

temperature

Temperature measured from the VPR CTD in celsius

salinity

Salinity measured from the VPR CTD

fluor_ref

A reference fluorescence baseline provided in millivolts by the VPR CTD for calibrating fluorescence_mv data

fluorescence_mv

Fluorescence in millivolts from the VPR CTD (uncalibrated)

turbidity_ref

A reference turbidity baseline provided in millivolts for calibrating turbidity_mv

turbidity_mv

Turbidity in millivolts from the VPR CTD (uncalibrated)

altitude_NA

Altitude data from the VPR CTD

day

Day on which VPR data was collected (from AutoDeck)

hour

Hour during which VPR data was collected (from AutoDeck)

station

Station idnetifier provided during processing

sigmaT

Density caluclated from temperature, pressure and salinity data

depth

Depth in metres caluclated form pressure