VPR measurement data calculated by Visual Plankton

A data frame of measurement information for each ROI in the sample data set including long axis length, perimeter and area, produced by vpr_autoid_read

roimeas_dat_combine

Format

A data frame with 12 variables

roi

Unique ROI identifier - 10 digit

taxa

Category in which ROI has been classified by Visual Plankton

day_hour

day and hour in which data was collected (from Autodeck)

Perimeter

The perimeter of the ROI in millimeters

Area

The area of the ROI in millimeters

width1

Width at a first point of the ROI in millimetres (defined in more detail in VPR manual)

width2

Width at a second point of the ROI in millimetres (defined in more detail in VPR manual)

width3

Width at a third point of the ROI in millimetres (defined in more detail in VPR manual)

short_axis_length

The length in millimeters of the ROI along the shorter axis

long_axis_length

The length in millimeters of the ROI along the longer axis

station

Station identifier provided in processing

time_ms

Time stamp when ROI was collected in milliseconds