VPR size information dataframe

A sample data frame of size information from Visual Plankton outputs, processed using vpr_ctdroisize_merge

size_df_f

Format

An object of class data.frame with 14 rows and 14 columns.

Details

@format A dataframe with 14 variables including

frame_ID

Unique identifier for each VPR frame

pressure

Pressure measured from the VPR CTD in decibars

temperature

Temperature measured from the VPR CTD in celsius

salinity

Salinity measured from the VPR CTD

sigmaT

Density calculated from temperature, salinity and pressure

fluorescence_mv

Fluorescence measured by the VPR CTD in millivolts (uncalibrated)

turbidity_mv

Turbidity measured by the VPR CTD in millivolts (uncalibrated)

roi

Unique ROI identification number - 10 digits, 8 digit millisecond time stamp and two unique digits to denote multiple ROIs within a millisecond

taxa

Category in which ROI has been classified by Visual Plankton

day_hour

Day and hour in which data was collected, from AutoDeck processing

long_axis_length

The length of the longest axis of the ROI image, measured by Visual Plankton

station

Station identifier provided during processing

time_ms

Time stamp when ROI was collected (milliseconds)

roi_ID

ROI identification number- 8 digit time stamp, without unique 2 digit ending