A binned data frame of concentration data per category

A 'binned' dataframe from sample VPR data, including concentrations of each category, where each data point represents a 5 metre bin of averaged VPR data. Produced using vpr_roi_concentration

taxa_conc_n

Format

A dataframe with 21 variables

depth

Depth calculated from pressure in metres

min_depth

The minimum depth of the bin in metres

max_depth

The maximum depth of the bin in metres

depth_diff

The difference between minimum and maximum bin depth in metres

min_time_s

The minimum time in seconds of the bin

max_time_s

The maximum time in seconds of the bin

time_diff_s

The difference between minimum and maximum time in a bin, in seconds

n_roi_bin

The number of ROI observations in a bin

conc_m3

The concentration of ROIs in a bin, calculated based on image volume and number of frames per bin

temperature

Temperature measured from the VPR CTD in celsius (averaged within the bin)

salinity

Salinity measured from the VPR CTD (averaged within the bin)

density

sigma T density calculated from temperature, salinity and pressure (averaged within the bin)

fluorescence

Fluorescence measured by the VPR CTD in millivolts (uncalibrated) (averaged within the bin)

turbidity

Turbidity measured by the VPR CTD in millivolts (uncalibrated) (averaged within the bin)

avg_hr

The mean time in which bin data was collected, in hours

n_frames

The number of frames captured within a bin

vol_sampled_bin_m3

The volume of the bin sampled in metres cubed

toyo

Identifier of the tow-yo section which bin is a part of, either ascending or descending, appended by a number

max_cast_depth

The maximum depth of the entire VPR cast

taxa

The category in which ROIs in bin have been classified by Visual Plankton

station

Station identifier provided during processing