Checks manually created aid files for errors

Removes any empty aid files after manual reclassification, checks for tow numbers and other metadata to match. Performs check to ensure measurement and ROI files are the same length

vpr_autoid_check(basepath, cruise, del)

Arguments

basepath

basepath to autoid folder eg. C:/data/CRUISENAME/autoid/

cruise

name of cruise which is being checked

del

Logical value, if TRUE, empty files will be deleted (see warning), if FALSE, files WILL NOT be deleted (they will be listed in output)

Value

text file (saved in working directory) named CRUISENAME_aid_file_check.txt

Details

WARNING: This function will delete empty aid and aidmeas files, permanently changing your directory. Consider making a back up copy before running this function.

Author

E Chisholm