Modifies aid and aid mea files based on manual reclassification

Modifies aid and aid mea files based on manual reclassification

vpr_autoid_create(
  reclassify,
  misclassified,
  basepath,
  day,
  hour,
  mea = TRUE,
  categories
)

Arguments

reclassify

list of reclassify files (output from vpr_manual_classification())

misclassified

list misclassify files (output from vpr_manual_classification())

basepath

base path to auto ID folder eg 'E:/autoID_EC_07032019/'

day

day identifier for relevant aid & aidmeas files

hour

hour identifier for relevant aid & aidmeas files

mea

logical indicating whether or not there are accompanying measurement files to be created

categories

A list object with all the potential classification categories

Author

E. Chisholm

Examples

if (FALSE) { # \dontrun{
basepath <- 'E:/autoID_EC_07032019/'
day <- '289'
hr <- '08'
categories <-
c("bad_image_blurry", "bad_image_malfunction", "bad_image_strobe", "Calanus", "chaetognaths",
"ctenophores", "krill", "marine_snow", "Other", "small_copepod", "stick")
day_hour_files <-  paste0('d', day, '.h', hr)
misclassified <- list.files(day_hour_files, pattern = 'misclassified_', full.names = TRUE)
reclassify <- list.files(day_hour_files, pattern = 'reclassify_', full.names = TRUE)
vpr_autoid_create(reclassify, misclassified, basepath, categories)
} # }