Function to check results of classification manually

Displays each image in day hour specified, prompts user to confirm or deny classification. If classification is denied, asks for a reclassification value based on available taxa

vpr_manual_classification(
  day,
  hour,
  basepath,
  taxa_of_interest,
  gr = TRUE,
  scale = "x300",
  opticalSetting = "S2",
  img_bright = TRUE
)

Arguments

day

day of interest in autoid

hour

hour of interest in autoid

basepath

file path to auto id folder eg 'E:/autoID_EC_07032019/'

taxa_of_interest

list of taxa folders you wish you sort through

gr

logical indicating whether pop up graphic menus are used (user preference - defaults to TRUE)

scale

argument passed to image_scale, default = 'x300'

opticalSetting

specifies optical setting of VPR, defining image frame size, current options are 'S0', 'S1', 'S2' (default), 'S3', see further info in details

img_bright

logical value indicating whether or not to include a blown out high brightness version of image (can be helpful for viewing dark field fine appendages)

Details

Optical Setting frame sizes: S0 = 7x7 mm, S1 = 14x14mm, S2 = 24x24mm, S3 = 48x48 mm. These settings define the conversion factor from pixels to millimetres and calculate image size for classification reference

Development

Add "undo" functionality to go back on a typing mistake

Fix scaling/ size issue so images are consistently sized

show ROI number for image somewhere for reference when in doubt of classification