Bin VPR size data

Calculates statistics for VPR measurement data in depth averaged bins for analysis and visualization

vpr_size_bin(data_all, bin_mea)

Arguments

data_all

a VPR CTD and measurement dataframe from vpr_ctdroisize_merge

bin_mea

Numerical value representing size of depth bins over which data will be combined, unit is metres, typical values range from 1 - 5

Value

a dataframe of binned VPR size data statistics including number of observations, median, interquartile ranges, salinity and pressure, useful for making boxplots

Examples

data('size_df_f')
vpr_size_bin(size_df_f, bin_mea = 5)
#> Warning: data length is not a multiple of split variable
#> Warning: data length is not a multiple of split variable
#>      median   IQR1.1 IQR1.2 IQR1.3 IQR1.4   IQR1.5 IQR1.6 IQR1.7 IQR1.8 IQR1.9
#> 1  2.022656 2.022656     NA     NA     NA 2.347031     NA     NA     NA     NA
#> 2        NA 2.022656     NA     NA     NA 2.347031     NA     NA     NA     NA
#> 3        NA 2.022656     NA     NA     NA 4.852852     NA     NA     NA     NA
#> 4        NA 2.022656     NA     NA     NA 7.358672     NA     NA     NA     NA
#> 5  4.852852 2.022656     NA     NA     NA 7.358672     NA     NA     NA     NA
#> 6        NA 2.022656     NA     NA     NA 2.347031     NA     NA     NA     NA
#> 7        NA 2.022656     NA     NA     NA 2.347031     NA     NA     NA     NA
#> 8        NA 2.022656     NA     NA     NA 4.852852     NA     NA     NA     NA
#> 9        NA 2.022656     NA     NA     NA 7.358672     NA     NA     NA     NA
#> 10 2.967656 2.022656     NA     NA     NA 7.358672     NA     NA     NA     NA
#>     IQR1.10   IQR3.1 IQR3.2 IQR3.3 IQR3.4   IQR3.5 IQR3.6 IQR3.7 IQR3.8 IQR3.9
#> 1  2.266641 2.022656     NA     NA     NA 2.347031     NA     NA     NA     NA
#> 2  2.441895 2.022656     NA     NA     NA 2.347031     NA     NA     NA     NA
#> 3  2.967656 2.022656     NA     NA     NA 4.852852     NA     NA     NA     NA
#> 4  2.989805 2.022656     NA     NA     NA 7.358672     NA     NA     NA     NA
#> 5  2.997187 2.022656     NA     NA     NA 7.358672     NA     NA     NA     NA
#> 6  2.266641 2.022656     NA     NA     NA 2.347031     NA     NA     NA     NA
#> 7  2.441895 2.022656     NA     NA     NA 2.347031     NA     NA     NA     NA
#> 8  2.967656 2.022656     NA     NA     NA 4.852852     NA     NA     NA     NA
#> 9  2.989805 2.022656     NA     NA     NA 7.358672     NA     NA     NA     NA
#> 10 2.997187 2.022656     NA     NA     NA 7.358672     NA     NA     NA     NA
#>     IQR3.10 n_obs temperature salinity pressure
#> 1  2.266641     2  16.8294000 30.70469      5.5
#> 2  2.441895     0          NA       NA     10.5
#> 3  2.967656     0          NA       NA     15.5
#> 4  2.989805     0          NA       NA     20.5
#> 5  2.997187     4   3.5086500 30.75547     25.5
#> 6  2.266641     0          NA       NA     30.5
#> 7  2.441895     0          NA       NA     35.5
#> 8  2.967656     0          NA       NA     40.5
#> 9  2.989805     0          NA       NA     45.5
#> 10 2.997187     6   0.4229333 30.71310     50.5